secret-note-python
This program I wrote as a quick and easy note taking program.  By simply clicking the Python file icon, the terminal pops up prompting the user to make an entry.  As soon as the user presses enter, the terminal disappears and the entry is added to a file in the program file's directory.  The text file will be created by the program if one doesn't already exist.  The entry is added to the file with a date and time stamp and formatted to fit within a 96 character limit per line in the .txt file, as to prevent text going far off the page, and also spaced out 1 line from the time and date stamp, and two additional lines after each entry for easy reading.  

A shortcut icon on the desktop works well to make easy use of this program for quick note taking.